refactor(lark): separate Goal Topic routing rules
